Added option to BeastDialog to show list of BEAGLE resources and stop. Added 'Automatic' thread pool option. With BEAGLE, this should be the best option if multiple resources are being used.
Mac ANT release builder now includes the standard BEAGLE library location in java.library.path (/usr/local/lib) so if it is installed there it should be found. Note that in order to use the current BEAGLE with NVidia Cuda, the application must be run in 32-bit mode (use Get Info in Finder).

git-svn-id: http://beast-mcmc.googlecode.com/svn/trunk@2367 ca793f91-a31e-0410-b540-2769d408b6a1