Implemented some microtiming in CompoundLikelihood. This is turned on by a static constant in this class so won't be compiled unless switched on. The results can be obtained by dumping the CompoundLikelihood into a <report> element:
<report>
    <posterior idref="posterior"/>
</report>

This will break down the total number of evalations, the total amount of time and the time per eval for every prior and likelihood. This will be useful for considering parallelization and threading options.

The switch (CompoundLikelihood.EVALUATION_TIMERS) is on at present as I don't believe it will have any performance implications but I will be checking this.

git-svn-id: http://beast-mcmc.googlecode.com/svn/trunk@3447 ca793f91-a31e-0410-b540-2769d408b6a1