Revise naming of "effects".
Instead of "serum effect", we have "serum potency" and instead 
of "virus effect", we have "virus avidity".  This is to move 
things in better agreement with the literature.  Also, I want to 
distinguish serum potency from serum breadth, which is easier 
with proper language. Andrew, let me know if you don't agree 
with this change.


git-svn-id: http://beast-mcmc.googlecode.com/svn/trunk@6026 ca793f91-a31e-0410-b540-2769d408b6a1